+++ controllers/beatmap_set_route.hh
@@ -4,7 +4,7 @@
 using namespace drogon;
 class beatmap_set_route : public drogon::HttpController<beatmap_set_route> {
 public:
-    Task<HttpResponsePtr> get(HttpRequestPtr req, int32_t id);
+    Task<HttpResponsePtr> get(HttpRequestPtr req, int64_t id);
 
     METHOD_LIST_BEGIN
         ADD_METHOD_TO(beatmap_set_route::get, "/s/{1}", Get);
